Laravel event listeners are now used for updating NoT badges
"NoT" badges are badge-like numbers beside the note names on the note card.
Laravel event listeners are now being used for updating these in such a way
that payloads have become much smaller. Instead of broadcasting an array of
notes matching the structure of Vue's reactive "notes" array, only an object
containing the note ID and the updated number(s) are passed in the broadcast
message.

Currently, only the number of pending tasks is being shown on the web page.
The number of done tasks will be added later.